[ci] remove old USER_LAND code from ci scripts
We used to have a parser in jenkins that looked for "ENTERED USER LAND"
and "EXITED USER LAND" as a coarse way to classify failures in to infra
vs. user failures.

This is a good idea and I think we should implement it again for GHA.
But at the moment we don't have it, so remove the unused code.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/80045

Approved by: https://github.com/janeyx99