[Proof of Concept] Use labels to select the test configs to run (#83690)
This is the proof-of-concept PR to support linux.  Other platforms will follow in subsequent PRs.  Per feedbacks from the team, I have changed the label to be `test-config/CONFIG`, for example `test-config/functorch` to make it clear that this is not `ciflow`.

* The script maintains a set of valid test configs (shard names) including `default`, `functorch`, `dynamo`, etc.
* If the PR has one or more labels as specified in the set, i.e. **test-config/functorch**, only these test configs will be selected. If the PR has both `test-config/functorch` and `ciflow/trunk`, both will be taken into account: **All functorch builds and tests in trunk will be run**
* If the PR has none of the test-config label, all tests are run as usual.

Basically, the CI workflow will be `filter (part of build) -> build -> filter -> test[filtered_matrix]`. The filter is applied twice before build and test because we want to get the latest labels from the PR right before the steps are run. This is mainly to avoid GHA static list of labels that is only populated at the time of the pull request event, for example, a new pull request will have no label,

### Testing
This PR has a bunch of random labels but it includes two important labels among them `test-config/functorch` and `test-config/dynamo`.  The former was added before the CI started while the latter was added after (but before the test started).  Only functorch and dynamo tests (multiple shards) were run.

Also, I manage to find a way to hide the majority of skipped tests, so they won't clutter the signal box that much

### Issues
#82367
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83690
Approved by: https://github.com/ZainRizvi